renamed actors packages to reports to reflect their new content
