+++ gcm4/src/main/java/plugins/materials/reports/StageReport.java
@@ -1,4 +1,4 @@
-package plugins.materials.actors;
+package plugins.materials.reports;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
@@ -12,8 +12,8 @@ import plugins.materials.events.StageOfferUpdateEvent;
 import plugins.materials.support.MaterialsProducerId;
 import plugins.materials.support.StageId;
 import plugins.reports.support.ReportHeader;
-import plugins.reports.support.ReportLabel;
 import plugins.reports.support.ReportItem;
+import plugins.reports.support.ReportLabel;
 
 /**
  * A Report that displays the creation, destruction, offering, batch conversion,
