+++ gcm4/src/test/java/plugins/personproperties/reports/AT_PersonPropertyReport.java
@@ -1,4 +1,4 @@
-package plugins.personproperties.actors;
+package plugins.personproperties.reports;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
