+++ gcm4/src/test/java/plugins/reports/support/AT_PeriodicReport.java
@@ -10,8 +10,8 @@ import java.util.List;
 import org.apache.commons.math3.util.FastMath;
 import org.junit.jupiter.api.Test;
 
-import nucleus.ReportContext;
 import nucleus.Plugin;
+import nucleus.ReportContext;
 import nucleus.Simulation;
 import nucleus.testsupport.testplugin.TestActorPlan;
 import nucleus.testsupport.testplugin.TestPlugin;
