+++ gcm4/src/test/java/plugins/resources/reports/AT_ResourceReport.java
@@ -1,4 +1,4 @@
-package plugins.resources.actors;
+package plugins.resources.reports;
 
 import org.junit.jupiter.api.Test;
