+++ gcm4/src/test/java/plugins/util/properties/AT_IntPropertyManager.java
@@ -13,11 +13,11 @@ import org.junit.jupiter.api.Test;
 
 import nucleus.DataManagerContext;
 import nucleus.SimulationContext;
-import nucleus.testsupport.testplugin.TestPluginFactory;
-import nucleus.testsupport.testplugin.TestSimulation;
 import nucleus.testsupport.testplugin.TestActorPlan;
 import nucleus.testsupport.testplugin.TestDataManager;
 import nucleus.testsupport.testplugin.TestPluginData;
+import nucleus.testsupport.testplugin.TestPluginFactory;
+import nucleus.testsupport.testplugin.TestSimulation;
 import util.annotations.UnitTestConstructor;
 import util.annotations.UnitTestMethod;
 import util.errors.ContractException;
