+++ tutorials/plugins/gcm4_lesson_15_regions_plugin/src/main/java/lesson/Example_15.java
@@ -21,8 +21,8 @@ import plugins.people.PeoplePluginData;
 import plugins.people.support.PersonId;
 import plugins.regions.RegionsPlugin;
 import plugins.regions.RegionsPluginData;
-import plugins.regions.actors.RegionPropertyReport;
-import plugins.regions.actors.RegionTransferReport;
+import plugins.regions.reports.RegionPropertyReport;
+import plugins.regions.reports.RegionTransferReport;
 import plugins.reports.ReportsPlugin;
 import plugins.reports.ReportsPluginData;
 import plugins.reports.support.NIOReportItemHandler;
