+++ tutorials/plugins/gcm4_lesson_17_groups_plugin/src/main/java/lesson/Example_17.java
@@ -26,12 +26,12 @@ import plugins.globalproperties.GlobalPropertiesPluginData.Builder;
 import plugins.globalproperties.support.GlobalPropertyId;
 import plugins.groups.GroupsPlugin;
 import plugins.groups.GroupsPluginData;
-import plugins.groups.actors.GroupPopulationReport;
+import plugins.groups.reports.GroupPopulationReport;
 import plugins.people.PeoplePlugin;
 import plugins.people.PeoplePluginData;
 import plugins.personproperties.PersonPropertiesPlugin;
 import plugins.personproperties.PersonPropertiesPluginData;
-import plugins.personproperties.actors.PersonPropertyReport;
+import plugins.personproperties.reports.PersonPropertyReport;
 import plugins.regions.RegionsPlugin;
 import plugins.regions.RegionsPluginData;
 import plugins.reports.ReportsPlugin;
